requirements/base.txt
beautifulsoup4==4.6.0
chardet==3.0.4
coreapi==2.3.3
defusedxml==0.5.0
diskcache==3.1.1
Django>=1.11,<2.0
django-allauth==0.39.1
django-appconf==1.0.2
django-braces==1.13.0
django-compressor==2.2
django-cache-machine==1.1.0
django-contrib-comments==1.9.1
django-extensions==2.1.6
django-filter==2.1.0
django-formtools==2.1
django-gravatar2==1.4.2
django-haystack==2.8.1
django-htmlmin==0.10.0
django-model-utils==3.1.2
django-pandas==0.6.1
django-rest-auth==0.9.3
django-rest-swagger==2.2.0
django-simple-captcha==0.5.10
django-tagging==0.4.6
django-uuslug==1.1.8
django-environ==0.4.5
django-webpack-loader==0.6
djangorestframework==3.9.3
djangorestframework-jwt==1.11
django-editormd>=0.3
django-oauth-toolkit==1.1.3
oauthlib==2.1.0

django-anymail==6.0


docopt==0.6.2
enum34==1.1.6
funcsigs==1.0.2
html5lib==1.0.1
ipaddress==1.0.18
jieba==0.39
Jinja2==2.10.1
Markdown==2.6.11
MarkupSafe==1.0
numpy==1.16.1
olefile==0.44
openapi-codec==1.3.2
pandas==0.24.1
Pillow==5.4.1
prompt-toolkit>=1.0.15
pyasn1==0.2.3
pycodestyle==2.3.1
pycparser==2.17
pygal==2.4.0
Pygments==2.3.1
pygments-markdown-lexer==0.1.0.dev39
PyJWT==1.7.1
pymdown-extensions==4.9.1
python-dateutil==2.8.0
python-openid==2.2.5
python-slugify==1.2.4
python3-openid==3.1.0
urllib3<1.25,>=1.21.1
requests==2.21.0
requests-oauthlib==0.8.0
simplejson==3.15.0
sqlparse==0.2.1
uritemplate==3.0.0
wcwidth==0.1.7
webencodings==0.5.1
xmltodict==0.11.0


#
# install from github
# ----------------------------------------------------------------------------------------------------------------------
-e git+https://github.com/edison7500/django-bulma.git#egg=django_bulma
